requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
PyMuPDF==1.25.5
pytesseract==0.3.13
Werkzeug==3.1.3
